Add hooks for form handling
We really only need to include the VIP submission form with VIP sites,
so let's add a couple hooks that let us add forms and add the VIP form
using those hooks. The VIP form now lives in vip-scanner-wpcom and
only shows if the VIP Theme Review is being used to scan the theme.