Cache the scanner results to speed up exports
Still run a scan every time we refresh the page. If the results are
different than what's in the cache, update it. This is nice because we
can use the cache to power exports. Instead of rescanning the theme, we
can use what's in the transient.

This greatly speeds up exports