Temporarily disabled email submission
Sending emails from localhost isn't going to work. It will probably work
in VIP Quickstart because the SMTP server is configured properly, but
in normal MAMP environments, it almost surely won't. Since we don't
know whether emails went through or not, we can't tell the user
if their submission failed.

We can turn this back on when this component is merged into Quickstart.