Rework the conditional 'error summary' field
Use the VIP_Scanner_Form::add_field() api instead and make it optional.

Currently there is no good way to do conditionally required fields in
VIP_Scanner_Form. The approach here was causing all fields to be
emptied out on an invalid POST, since it reset the transient.

Since there is no good way to do this currently (I struggled against it
for several hours), we can just make it optional for now and ding
anyone that fails to provide an explanation.