Suppress errors for transients that are too big
We're using transients to cache the review results, but if there are too many errors, the transient might be too big to fit into a 1MB cache bucket. It won't get added to the cache, which isn't the end of the world — it will just be slower.

We should still suppress the error telling us that the object is too big for the cache though.