Integrating PHP Code Sniffer.
Created a `WordPressCodingStandardsCheck` class that checks the theme
against the WordPress Coding Standards available here:
https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards

For `WordPressCodingStandardsCheck` to work PHP CodeSniffer must be
intalled and available in the shell as `phpcs`. The WordPress coding
standard must also be installed and available to `phpcs` as `WordPress`.
Instructions for this can be found on the GitHub page.

The `WordPressCodingStandardsCheck` was added only to the `WP.com Theme
Review` review.

The check parses the severity of the error from `phpcs` and tries to match
it with a known error level. Errors are printed out with the error text
from `phpcs` along with the file and the line that caused the error.

See: #54