Added checking for error slugs and classes to WordPressCodingStandardsCheck.php.
When reporting errors the `WordPressCodingStandardsCheck` now searches
through a list of known error slugs and slug classes to find an
appropriate error level for the error. If the error is not known then its
level will be a warning if the CodeSniffer level was an Error and a note
otherwise.

Specific error slugs can be added to the $check_level array. This setting
has first priority for setting the error level.

Error classes can be defined within the $check_class_level array. A class
definition of `WordPress.XSS.` would match any error slug that starts with
`WordPress.XSS.`. Classes have second priority for setting error level.

If an error slug was not known and its priority had to be computed by
searching through error classes and/or the reported level from
CodeSniffer, then its computed level will be cached for future searches.

The default error slug levels and class levels were derived using the
following error level scheme:

- `Blockers`: Most likely a severe security or performance concern.
- `Warnings`: Most likely a performance or other concern.
- `Notes`: Stylistic issues or other issues that do not effect security or
  performance.