Adding class to retrieve information about a PHP files' class, functions etc...
The `AnalyzedFile` class analyzes a file for PHP constructs such as namespaces, classes, functions, etc.. when instantiated. This class will be used for doing theme analysis.

This class uses mostly regular expresssions under the hood to do the parsing. While tokens could've been used instead, using regular expressions allows us to do less parsing of the exact contents of functions and the like at the cost of a bit of obscurity.