Adding Analyzer metas.
Analyzer metas are designed to arrange and present meta information about the theme and files that does not reflect the result of a check or any part of the review.

There is a meta class for the different things we need to track meta information for. This includes things like files and functions. Each classes is customized slightly to best present its' information.

The `MetaGroup` class is a special class because it is designed to act as a container rather than a meta for a specific thing. It can be used to group similar metadata elements together.

After an Analyzer is run it should populate a number of meta items representing the data it has aggregated. These items will eventually be surfaced based on their hierarchy. (Each meta can be given child elements, and a child may belong to more than one parent.