Customizing the scanner classes to allow for specifications of analyzers for individual reviews.
Customizing the scanner classes so that we may now specify which things should be analyzed by a review, similarily to how we already specify which checks to run.

We now pass `register_review()` an optional second array which contains a list of Analyzers to run, in the same format as checks are already passed. Both the checks and the analyzers eventually get passed down to the `BaseScanner` class.

The `BaseScanner::scan()` function now takes an optional argument `$scanners` argument which specifies which scanners to run. By default it will run both `checks` and `analyzers` but changing the values in this array will change what is run.