Adding WP CLI support for analyzers.
Analyzers can now be run from WP-CLI by running `wp vip-scanner analyze-theme`. This command is mostly the same as `scan-theme` except it takes a `--depth` parameter for how much of the hierarchy should be printed out.

To make it easier to write renderer code that will display nicely both via the web UI and the CLI I added the new `AnalyzerRenderer::stylize_text()` function. When the text is being renderered for web it will contain class names and proper html elements. When text is being rendered for the CLI it will instead contain bash text modifier codes and class names will be removed. All renderer classes were updated to output using this function.