Stopping showing of redundant or extraneous attributes in analyzer output.
Alot of information for the renderer is stored as an attribute. Unfortunately this causes some things to be displayed within the attributes when they do not need to be. Firstly, we check if the attribute is empty, and then we check if its in a blacklist of attributes that should not be shown.

The FileRenderer also re-defines this method so that the file line count is not shown, as it already appears in the header for that element.