Adding version number to cached review transient.
When we upgrade the VIP Scanner version number, we shouldn't be trying to
load old cached theme reviews
(https://github.com/Automattic/vip-scanner/blob/master/vip-scanner.php#L430).

This is because the structure of the `VIP_Scanner` or any of the other
objects we have may have changed - which can cause a fatal error if the
cached transient differs from the current one.

Adding the version number to the transient key ensures that if we do a
version upgrade old reviews are no longer loaded.