Adding better checker for relative paths in `ThemeScanner`.
Previously the `ThemeScanner` was only checking for the presence of a '/'
in the theme to determine if it was a path or a theme name. This could
fail if the desired theme was within a subfolder of the `themes` folder
other than `vip` or `pub`.

Now, we check and see if the relative path exists and if it does we assume
the user supplied a path rather than a theme name. This will also allow
specifying relative paths such as `.` or `../../`.

See bd523ad7621046a4ad6ae8cb58e9535e96f71536