Moved check logic to new method to be able to write more thorough unit tests
Many of the checks introduced in the VIP parameters check could not be exercised since the only checks in place so far were for the first parameter of a single parameter function. By moving the checking logic to a separate method that received the file path, file contents and list of checks as arguments, more thorough unit tests can be written.