Migrate VIPParametersTest to CodeCheckTestBase, dropping pointless tests
It doesn't make any sense to create new kinds of checks on the fly
at test scope, especially not in the way this was done here. There
was e.g. a check that warned about the presence of a numeric parameter
value but only checked for the number 5 (which it was given at test
scope, thus passing the test).