VIP Scanner: Added check for PHP variable variables.
VIPRestrictedPatternsCheck.php was using preg_match() and would not show a complete list of found results, so I created a new Check that uses preg_match_all() to list all of the results.

Uses the following regex: ((?<![\'\"])\$\$[a-zA-Z_\x7f-\xff][a-zA-Z0-9_\x7f-\xff]*?(?![\'\"])|(?<![\'\"])\$\{(?:.*)[\}](?![\'\"]))

If not preceeded by either " or ', search for "$$" and a valid PHP variable name, as long as it is not followed by either " or '
OR
If not preceeded by either " or ', search for "${" and end the search with "}", as long as it is not followed by either " or '

Try as I might, I was unable to find a solution to some false positives using regex.

If a file has embedded content outside of PHP tags, the content may still be caught by the regex.  Example:

<?php get_header(); ?>
<h1>Government wasted $$$$millions; What next?</h1>
<?php ...

We cannot make another check to verify that the regex IS inside proper PHP tags because there may not be a closing tag on the file.
This approach could cause the regex to fail to find actual potential variable variables.