VIP Scanner: The regular expression has now been modified to match variable class properties.
Example:
$foo->$bar

The primary section that does the matching is: \$[a-zA-Z_\x7f-\xff][a-zA-Z0-9_\x7f-\xff]*?\-\>\$

It will match against a properly named PHP variable, an object operator (->), and a dollar sign ($).

There should not be any need to match against complex syntax variable variables as class properties ($foo->${$bar}) because those should already be caught by the rest of the regular expression.