VIP Scanner: Modified VIPRestrictedPatternsCheck to allow for variable variable searches.
Bringing the variable variable search back into VIPRestrictedPatternsCheck will reduce the number of classes that have to iterate over the theme files.
By changing check() to use preg_match_all it shows all of the matches of a complex regex.
With this change, we do not need VIPRestrictedVariablesCheck so the class file can be removed and the reference can be taken out of config-vip-scanner.php.