VIP Scanner: Added basic search for hardcoded 'http' protocol in CSS files.
Per GitHub issue #216, in most instances protocol relative URLs should be used to avoid "mixed content" security warnings.

Right now we're only able to search CSS/JS files, but more checks can be added.  Instead of searching all PHP/HTML files for URLs to scan, it might be better to scan specific areas that cause trouble, such as script/style enqueues.

The primary problem that I see right now is that I do not see a way to check of a link is on the "same domain" since the VIP Scanner may be running on a domain that is not the final "live" domain.