VIP Scanner: Moved scan description escaping from display_theme_review_result_row() to check() in XSSVectorsCheck.
After working further, I feel that there may actually be cases where HTML is valid in the scan descriptions. If escaping is necessary, it is probably better to be done on a per-case basis before the description goes to display_theme_review_result_row().

Example: If one wants to link to more information about a security concern or best practice, it may be good to have an HTML link in the description.