Fix inconsistent caching of scan results
Change cache structure from storing theme path in post name to storing
in post meta instead. This is because WP_Query does not behave as
intended when dealing with duplicate post names (it assumes all post
name to be unique) and in this case would ignore taxonomy criteria if
there's a post name match, resulting in wrong cache returned.

This also addresses issue #144 and allows toolbar to display correct
count from cache as intended.