Check node type in function call checks
Before invoking `$node->name->toString()` on a node of type
`PhpParser\Node\Expr\FuncCall`, make sure that `$node->name`
really has a `toString()` member, by checking that it is of type
`PhpParser\Node\Name`.

In cases where the function invoked is specified by a variable
(`PhpParser\Node\Expr\Variable`, as in `$function();`)
instead of a name, the check would otherwise throw, as a variable
doesn't have a `toString()` member (and we can't compare that variable
to a blacklist of forbidden function names anyway).

Props to @fklein-lu for reporting this issue.