Flagging get_intermediate_image_sizes() in VIP, no intermediate images there.
Added check for get_intermediate_image_sizes() when running VIP Theme Review.
As there are no intermediate images on VIP, calls to this function are
'Blockers'. Unit-test is included for this check.

This should solve issue #290.

Note that this patch does alter one line of unrelated test; this is
due to a misplaced bracket.